RavenDB-17966 Corax - static indexes
